Progress? What progress?
So, I basically started over. I realised that it would be far more
computationally efficient, and it would be far more difficult for me to
screw up, if I did cells instead of vertices and just approximated force
by measuring the gradient of energy empirically. The primary reason for
this was that it makes it so that I can now confidently compare energy
to energy gradient, and thus do Goldberg criteria for backtracking line
searches. Or at least, I will be able to, once I get back to where I
was.