Use Maven, they said.
It would be easy, they said.

That sucked.
Maybe it's just me, or m2e, but this seems like a terrible package manager. If you download a .jar and put it on your build path, then java sees it and tells you what it does and doesn't have. With this, though, I would frequently change something and then change it back, or remove an unused import, or restart eclipse, and the state of my project would change. At one point, it was unable to resolve the imports from my own src folder! It also didn't help that the folks at GeoTools outright lied about what dependencies their project had. I guess I'll leave it like this since it works (for now), but oh my god I'm never using Maven again.