#researchflatearth
I've created Danseiji IX, a map that serves to maximise the efficacy of flat earth arguments by demonstrating that a map with the topology of most flat earth maps can exist without distorting Australia like it normally does.
Not that I can see it with stupid Antarctica in the way.