Triangles are my favorite shape
I totally redid the mesh system to use triangular Elements instead of rectangular Cells. The goal was to help it deal with the really weird geometries around the outside of an azimuthal plot. It didn't really work, and it caused some new problems to appear, but the tears look way better now than they did before. I think I need to work with it some more and change the topology further.