This render took longer than you can imagine.
It works! I added a lengthy bit of code to shrink the initial condition to a reasonable place before starting L-BFGS, and it's _way_ faster, and works with the old start condition. It's not perfect; Antarctica and that side of the map don't shrink properly for subtleties of the mesh shape that I don't fully understand. Because of that, I suspect I'll still need the slower gradient solution. However, that one dilated side can be removed by the automatic optimisation far more quickly and easily than the entire periphery, like before.