YES!
It's gorgeous! I did have to add a new feature, which is that instead of just filling notches that are one element deep, it repairs entire cuts. Basically, whenever its total tear gets to the desired length, it will find the tear that is the closest to being closed and close it slightly. It repeats this, enabling it to get rid of entire cuts, like the vestigial one that formed by New Zealand again.