It's like a cartogram, but better.
I finished Danseiji VI! I ended up changing the population data file, by blurring it less (the original purpose of the blurring was to help it converge, but I'm pretty sure the new exact gradient makes that a guarantee) to bring the baseline density down. It looks super cool, and I'm quite happy with it. Fiddling with the image blurriing took way too long, though.