Curse you, Big Map Projection!
I made a new figure to illustrate why I can't use the Airy--Kavrayskiy criterion by implementing the Airy--Kavrayskiy criterion. The reason is that it's concave down in all the wrong places. I strongly question whether it's really a good measure for the quality of a map projection. I mean, look at the thing that minimized it! Look at it! This had an E_K = .337. That's lower than any map projection in existence save Gyorffy's optimized polynomial ones!